Feat added default avatar picture. Added avatar into templates: programmer_detail.html, programmers_list.html. Moved cloudinary.py into services directory.
