+++ knownProgrammer/accounts/static/css/style.css
@@ -83,11 +83,15 @@
     color: grey;
 }
 
+.star-active {
+    color: white;
+}
+
 .star-label {
     display: inline-block;
-    font-size: 24px; /* Adjust as needed */
-    color: #FFD700; /* Gold color for stars */
-    margin-right: 5px; /* Adjust as needed */
+    font-size: 24px;
+    color: #FFD700;
+    margin-right: 5px;
 }
 
 .card-title {
