+++ knownProgrammer/accounts/templates/accounts/programmer_detail.html
@@ -20,6 +20,7 @@
     <div class="col-md-6 offset-md-3">
         <div class="card">
             <div class="card-header card-title text-center">
+                <img src="{{ programmer.user_id.avatar }}" alt="User Avatar">
                 <h3>{{ programmer.user_id.first_name }} {{ programmer.user_id.last_name }}</h3>
             </div>
             <div class="card-body">
